project module tested

+++ src/modules/project/project.service.ts
@@ -6,6 +6,7 @@ import { CreateProjectDto } from './dto/create-project.dto';
 import { Promise } from 'mongoose';
 import { ProjectResponse } from './dto/response-project.dto';
 import { UpdateProjectDto } from './dto/update-project.dto';
+import { DeleteProjectResponseDto } from './dto/delete-project.dto';
 
 @Injectable()
 export class ProjectService {
@@ -14,21 +15,22 @@ export class ProjectService {
         @InjectModel(Project.name) private projectModel: Model<Project> 
     ){}
 
+    private query = "name projectType githubFrontend githubBackend liveUrl video description technologies features"
 
-    async create (createProjectDto: CreateProjectDto): Promise<ProjectResponse[]>  {
-        const project = new this.projectModel(createProjectDto);
+    async create (userId: string, createProjectDto: CreateProjectDto): Promise<ProjectResponse[]>  {
+        const project = new this.projectModel({...createProjectDto, userId:userId});
         await project.save();
-        return await this.projectModel.find({userId: createProjectDto.userId}).exec()
+        return await this.projectModel.find({userId: userId}).select(this.query).exec()
     }
 
 
-    async update (id:string, updateProjectDto: UpdateProjectDto) : Promise<ProjectResponse[]>{
-        if(!Types.ObjectId.isValid(id)){
+    async update (id:string, userId: string,  updateProjectDto: UpdateProjectDto) : Promise<ProjectResponse[]>{
+        if(!Types.ObjectId.isValid(id) || !Types.ObjectId.isValid(userId)){
             throw new BadRequestException("Invalid project id")
         }
 
 
-        const updated = await this.projectModel.findOneAndUpdate({_id:id, userId:updateProjectDto.userId}, updateProjectDto, {new:true, runValidators:true}).exec()
+        const updated = await this.projectModel.findOneAndUpdate({_id:id, userId:userId}, updateProjectDto, {new:true, runValidators:true}).exec()
 
 
         if(!updated){
@@ -36,8 +38,7 @@ export class ProjectService {
         }
 
 
-        return await this.projectModel.find({userId: updateProjectDto.userId}).exec()
-
+        return await this.projectModel.find({userId: userId}).select(this.query).exec()
 
     }
 
@@ -48,14 +49,20 @@ export class ProjectService {
             throw new BadRequestException("Invalid project id")
         }
 
-        return await this.projectModel.find({userId: userId}).exec()
+        return await this.projectModel.find({userId: userId}).select(this.query).exec()
 
     }
 
 
-    async remove (id: string, userId:string): Promise<any> {
+    async remove (id: string, userId:string): Promise<DeleteProjectResponseDto> {
         if(!Types.ObjectId.isValid(userId) || !Types.ObjectId.isValid(userId)){
             throw new BadRequestException("Invalid project id or user id")
+        }
+
+        const isFound = await this.projectModel.findOne({_id: id, userId: userId}).select(this.query).exec();
+
+        if(!isFound){
+            throw new NotFoundException("There is no project")
         }
 
         const removed = await this.projectModel.deleteOne({_id: id, userId: userId}).exec()
@@ -65,8 +72,13 @@ export class ProjectService {
             throw new NotFoundException("No data deleted")
         }
 
+        const remaining = await this.projectModel.find({userId:userId}).select(this.query).exec();
 
-        return removed
+
+        return {
+            removed: isFound,
+            remaining,
+        }
 
 
     }
